Add comprehensive tests for image conversion and inspection functionality
- Implemented unit tests for inspection features in `test_inspection.py`, covering basic inspection, handling of non-existent stores, and scenarios without metadata.
- Created integration tests in `test_integration.py` to validate comprehensive scenarios including mixed image formats, custom metadata, and resizing functionality.
- Added tests for direct memory conversion in `test_memory_conversion.py`, ensuring correct handling of NCHW format and path structure.
- Developed metadata handling tests in `test_metadata.py` to verify preservation and correctness of metadata during conversion.
- Introduced performance tests in `test_performance.py` to assess conversion speed and memory usage during large dataset processing.
