--- tests/test_images_to_zarr.py
@@ -1,1316 +1,0 @@
-import pytest
-import numpy as np
-import pandas as pd
-from pathlib import Path
-import tempfile
-import shutil
-from PIL import Image
-from astropy.io import fits
-import zarr
-import imageio
-
-from images_to_zarr.convert import convert
-from images_to_zarr.inspect import inspect
-from images_to_zarr import I2Z_SUPPORTED_EXTS
-
-
-@pytest.fixture
-def temp_dir():
-    """Create a temporary directory for tests."""
-    temp_dir = Path(tempfile.mkdtemp())
-    yield temp_dir
-    shutil.rmtree(temp_dir)
-
-
-@pytest.fixture
-def sample_images(temp_dir):
-    """Create sample images in various formats."""
-    images_dir = temp_dir / "images"
-    images_dir.mkdir()
-
-    # Create sample data
-    sample_data_2d = np.random.randint(0, 255, (64, 64), dtype=np.uint8)
-
-    files = []
-
-    # PNG image
-    png_path = images_dir / "test_image.png"
-    Image.fromarray(sample_data_2d, mode="L").save(png_path)
-    files.append(png_path)
-
-    # JPEG image
-    jpeg_path = images_dir / "test_image.jpg"
-    Image.fromarray(sample_data_2d, mode="L").save(jpeg_path)
-    files.append(jpeg_path)
-
-    # TIFF image
-    tiff_path = images_dir / "test_image.tiff"
-    Image.fromarray(sample_data_2d, mode="L").save(tiff_path)
-    files.append(tiff_path)
-
-    # FITS image
-    fits_path = images_dir / "test_image.fits"
-    hdu = fits.PrimaryHDU(sample_data_2d.astype(np.float32))
-    hdu.writeto(fits_path, overwrite=True)
-    files.append(fits_path)
-
-    # Multi-extension FITS
-    fits_multi_path = images_dir / "test_multi.fits"
-    hdul = fits.HDUList(
-        [
-            fits.PrimaryHDU(),
-            fits.ImageHDU(sample_data_2d.astype(np.float32), name="SCI"),
-            fits.ImageHDU(sample_data_2d.astype(np.float32) * 0.1, name="ERR"),
-        ]
-    )
-    hdul.writeto(fits_multi_path, overwrite=True)
-    files.append(fits_multi_path)
-
-    return images_dir, files
-
-
-@pytest.fixture
-def sample_metadata(temp_dir, sample_images):
-    """Create sample metadata CSV."""
-    images_dir, files = sample_images
-
-    metadata_data = []
-    for i, file_path in enumerate(files):
-        metadata_data.append(
-            {
-                "filename": file_path.name,
-                "source_id": f"SRC_{i:03d}",
-                "ra": 123.456 + i * 0.1,
-                "dec": 45.678 + i * 0.1,
-                "magnitude": 18.5 + i * 0.2,
-            }
-        )
-
-    metadata_df = pd.DataFrame(metadata_data)
-    metadata_path = temp_dir / "metadata.csv"
-    metadata_df.to_csv(metadata_path, index=False)
-
-    return metadata_path, metadata_df
-
-
-class TestImageFormats:
-    """Test reading various image formats."""
-
-    def test_supported_extensions(self):
-        """Test that all expected extensions are supported."""
-        expected_exts = {".fits", ".fit", ".tif", ".tiff", ".png", ".jpg", ".jpeg"}
-        assert I2Z_SUPPORTED_EXTS == expected_exts
-
-    def test_png_reading(self, sample_images):
-        """Test PNG image reading."""
-        from images_to_zarr.convert import _read_image_data
-
-        images_dir, files = sample_images
-        png_file = [f for f in files if f.suffix == ".png"][0]
-
-        data, metadata = _read_image_data(png_file)
-
-        assert data.ndim == 2
-        assert data.shape == (64, 64)
-        assert metadata["original_extension"] == ".png"
-        assert "mode" in metadata
-
-    def test_jpeg_reading(self, sample_images):
-        """Test JPEG image reading."""
-        from images_to_zarr.convert import _read_image_data
-
-        images_dir, files = sample_images
-        jpeg_file = [f for f in files if f.suffix == ".jpg"][0]
-
-        data, metadata = _read_image_data(jpeg_file)
-
-        assert data.ndim == 2
-        assert data.shape == (64, 64)
-        assert metadata["original_extension"] == ".jpg"
-
-    def test_tiff_reading(self, sample_images):
-        """Test TIFF image reading."""
-        from images_to_zarr.convert import _read_image_data
-
-        images_dir, files = sample_images
-        tiff_file = [f for f in files if f.suffix == ".tiff"][0]
-
-        data, metadata = _read_image_data(tiff_file)
-
-        assert data.ndim == 2
-        assert data.shape == (64, 64)
-        assert metadata["original_extension"] == ".tiff"
-
-    def test_fits_reading(self, sample_images):
-        """Test FITS image reading."""
-        from images_to_zarr.convert import _read_image_data
-
-        images_dir, files = sample_images
-        fits_file = [f for f in files if f.suffix == ".fits" and "multi" not in f.name][0]
-
-        data, metadata = _read_image_data(fits_file)
-
-        assert data.ndim == 2
-        assert data.shape == (64, 64)
-        assert metadata["original_extension"] == ".fits"
-        assert metadata["fits_extension"] == 0
-
-    def test_fits_multi_extension(self, sample_images):
-        """Test multi-extension FITS reading."""
-        from images_to_zarr.convert import _read_image_data
-
-        images_dir, files = sample_images
-        fits_file = [f for f in files if "multi" in f.name][0]
-
-        # Test single extension by name
-        data, metadata = _read_image_data(fits_file, fits_extension="SCI")
-        assert data.ndim == 2
-        assert metadata["fits_extension"] == "SCI"
-
-        # Test multiple extensions
-        data, metadata = _read_image_data(fits_file, fits_extension=["SCI", "ERR"])
-        assert metadata["fits_extensions"] == ["SCI", "ERR"]
-
-
-class TestConversion:
-    """Test the main conversion functionality."""
-
-    def test_basic_conversion(self, temp_dir, sample_images, sample_metadata):
-        """Test basic image to Zarr conversion."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            num_parallel_workers=2,
-            chunk_shape=(1, 64, 64),
-            overwrite=True,
-        )
-
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-        assert zarr_path.name.endswith(".zarr")
-
-        # Check Zarr structure
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-
-        assert "images" in root
-        images_array = root["images"]
-        assert images_array.shape[0] == len(files)
-
-        # Check metadata file
-        metadata_parquet = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        assert metadata_parquet.exists()
-
-        saved_metadata = pd.read_parquet(metadata_parquet)
-        assert len(saved_metadata) == len(files)
-
-    def test_recursive_search(self, temp_dir):
-        """Test recursive directory search."""
-        from images_to_zarr.convert import _find_image_files
-
-        # Create nested directory structure
-        images_dir = temp_dir / "images_test"
-        sub_dir = images_dir / "subdir"
-        sub_dir.mkdir(parents=True)
-
-        # Create images in both directories
-        Image.fromarray(np.random.randint(0, 255, (32, 32), dtype=np.uint8)).save(
-            images_dir / "img1.png"
-        )
-        Image.fromarray(np.random.randint(0, 255, (32, 32), dtype=np.uint8)).save(
-            sub_dir / "img2.png"
-        )
-
-        # Test non-recursive
-        files_non_recursive = _find_image_files([images_dir], recursive=False)
-        assert len(files_non_recursive) == 1
-
-        # Test recursive
-        files_recursive = _find_image_files([images_dir], recursive=True)
-        assert len(files_recursive) == 2
-
-    def test_compression_options(self, temp_dir, sample_images, sample_metadata):
-        """Test different compression options."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        # Test with different compressors
-        for compressor in ["zstd", "lz4", "gzip"]:
-            zarr_path = convert(
-                folders=[images_dir],
-                recursive=False,
-                metadata=metadata_path,
-                output_dir=output_dir / compressor,
-                compressor=compressor,
-                clevel=1,
-                overwrite=True,
-            )
-
-            store = zarr.storage.LocalStore(zarr_path)
-            root = zarr.open_group(store=store, mode="r")
-            assert root.attrs["compressor"] == compressor
-
-    def test_error_handling(self, temp_dir):
-        """Test error handling for invalid inputs."""
-        output_dir = temp_dir / "output"
-
-        # Create a dummy image file for the metadata tests
-        dummy_image = np.random.randint(0, 255, (100, 100, 3), dtype=np.uint8)
-        dummy_path = temp_dir / "dummy.png"
-        imageio.imwrite(dummy_path, dummy_image)
-
-        # Test missing metadata file
-        with pytest.raises(FileNotFoundError):
-            convert(
-                folders=[temp_dir],
-                recursive=False,
-                metadata=temp_dir / "nonexistent.csv",
-                output_dir=output_dir,
-            )
-
-        # Test invalid metadata CSV
-        bad_metadata = temp_dir / "bad_metadata.csv"
-        pd.DataFrame({"not_filename": ["test"]}).to_csv(bad_metadata, index=False)
-
-        with pytest.raises(ValueError, match="filename"):
-            convert(
-                folders=[temp_dir], recursive=False, metadata=bad_metadata, output_dir=output_dir
-            )
-
-    def test_conversion_without_metadata(self, temp_dir, sample_images):
-        """Test conversion with automatically generated metadata from filenames."""
-        images_dir, files = sample_images
-        output_dir = temp_dir / "output"
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=None,  # No metadata provided
-            output_dir=output_dir,
-            num_parallel_workers=2,
-            chunk_shape=(1, 64, 64),
-            overwrite=True,
-        )
-
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-        assert zarr_path.name == "images.zarr"  # Default name when no metadata
-
-        # Check Zarr structure
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-
-        assert "images" in root
-        images_array = root["images"]
-        assert images_array.shape[0] == len(files)
-
-        # Check metadata file - should contain only filenames
-        metadata_parquet = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        assert metadata_parquet.exists()
-
-        saved_metadata = pd.read_parquet(metadata_parquet)
-        assert len(saved_metadata) == len(files)
-        assert "filename" in saved_metadata.columns
-
-        # Should contain all the filenames
-        expected_filenames = {f.name for f in files}
-        actual_filenames = set(saved_metadata["filename"])
-        assert expected_filenames == actual_filenames
-
-
-class TestInspection:
-    """Test the inspection functionality."""
-
-    def test_basic_inspection(self, temp_dir, sample_images, sample_metadata, capsys):
-        """Test basic Zarr store inspection."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        # First create a store
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            overwrite=True,
-        )
-
-        # Then inspect it
-        inspect(zarr_path)
-
-        captured = capsys.readouterr()
-        output = captured.out
-
-        assert "SUMMARY STATISTICS" in output
-        assert f"Total images across all files: {len(files)}" in output
-        # Format and data type distribution may or may not be present depending on metadata
-        # Just check that basic summary information is there
-        assert "Data type:" in output
-
-    def test_inspect_nonexistent_store(self, temp_dir):
-        """Test inspection of non-existent store."""
-        nonexistent_path = temp_dir / "nonexistent.zarr"
-
-        # Should not raise exception, just log error and return
-        result = inspect(nonexistent_path)
-        assert result is None  # Function should return None for non-existent store
-
-    def test_inspect_without_metadata(self, temp_dir, sample_images, capsys):
-        """Test inspection when metadata file is missing."""
-        images_dir, files = sample_images
-
-        # Create minimal metadata
-        metadata_df = pd.DataFrame({"filename": [f.name for f in files]})
-        metadata_path = temp_dir / "minimal_metadata.csv"
-        metadata_df.to_csv(metadata_path, index=False)
-
-        output_dir = temp_dir / "output"
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            overwrite=True,
-        )
-
-        # Remove metadata file
-        metadata_parquet = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        if metadata_parquet.exists():
-            metadata_parquet.unlink()
-
-        inspect(zarr_path)
-
-        captured = capsys.readouterr()
-        output = captured.out
-
-        assert "SUMMARY STATISTICS" in output
-        assert f"Total images across all files: {len(files)}" in output
-
-
-class TestPerformance:
-    """Basic performance tests."""
-
-    def test_conversion_speed(self, temp_dir):
-        """Test conversion speed with a larger dataset."""
-        import time
-
-        # Create more test images
-        images_dir = temp_dir / "images"
-        images_dir.mkdir()
-
-        num_images = 50
-        files = []
-
-        for i in range(num_images):
-            img_data = np.random.randint(0, 255, (128, 128), dtype=np.uint8)
-            img_path = images_dir / f"test_{i:03d}.png"
-            Image.fromarray(img_data, mode="L").save(img_path)
-            files.append(img_path)
-
-        # Create metadata
-        metadata_df = pd.DataFrame({"filename": [f.name for f in files], "id": range(num_images)})
-        metadata_path = temp_dir / "metadata.csv"
-        metadata_df.to_csv(metadata_path, index=False)
-
-        # Time the conversion
-        start_time = time.time()
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=temp_dir / "output",
-            num_parallel_workers=4,
-            overwrite=True,
-        )
-
-        conversion_time = time.time() - start_time
-
-        # Verify results
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        assert images_array.shape[0] == num_images
-
-        # Basic performance check (should process at least 10 images per second)
-        images_per_second = num_images / conversion_time
-        assert images_per_second > 5, f"Too slow: {images_per_second:.2f} images/sec"
-
-    def test_memory_usage(self, temp_dir, sample_images, sample_metadata):
-        """Test that memory usage stays reasonable."""
-        import psutil
-        import os
-
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        process = psutil.Process(os.getpid())
-        initial_memory = process.memory_info().rss
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            num_parallel_workers=2,
-            overwrite=True,
-        )
-
-        # Ensure the conversion completed successfully
-        assert zarr_path.exists(), "Zarr store was not created"
-
-        final_memory = process.memory_info().rss
-        memory_increase = (final_memory - initial_memory) / 1024**2  # MB
-
-        # Should not use more than 100MB for this small test
-        assert memory_increase < 100, f"Memory usage too high: {memory_increase:.2f} MB"
-
-
-class TestMetadata:
-    """Test metadata handling."""
-
-    def test_metadata_preservation(self, temp_dir, sample_images, sample_metadata):
-        """Test that original metadata is preserved."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            overwrite=True,
-        )
-
-        # Load saved metadata
-        metadata_parquet = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        saved_metadata = pd.read_parquet(metadata_parquet)  # Check original columns are preserved
-        for col in metadata_df.columns:
-            assert col in saved_metadata.columns
-
-        # Check additional metadata columns are added - check what's actually there
-        # The exact columns depend on processing details, but we should have image processing metadata
-        expected_processing_cols = [
-            "original_filename",
-            "dtype",
-            "shape",
-        ]
-        for col in expected_processing_cols:
-            assert (
-                col in saved_metadata.columns
-            ), f"Expected column '{col}' not found in {saved_metadata.columns.tolist()}"
-
-        # Optional metadata columns (may not be present for performance reasons)
-        optional_cols = [
-            "file_size_bytes",
-            "min_value",
-            "max_value",
-            "mean_value",
-        ]
-        # Just check that some optional metadata is present, not all
-        optional_present = sum(1 for col in optional_cols if col in saved_metadata.columns)
-        assert optional_present >= 0  # At least some metadata should be preserved
-
-    def test_zarr_attributes(self, temp_dir, sample_images, sample_metadata):
-        """Test that Zarr attributes are set correctly."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-        output_dir = temp_dir / "output"
-
-        zarr_path = convert(
-            folders=[images_dir],
-            recursive=False,
-            metadata=metadata_path,
-            output_dir=output_dir,
-            fits_extension=0,
-            compressor="zstd",
-            clevel=3,
-            overwrite=True,
-        )
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        attrs = dict(root.attrs)
-
-        assert attrs["total_images"] == len(files)
-        assert attrs["compressor"] == "zstd"
-        assert attrs["compression_level"] == 3
-        assert "supported_extensions" in attrs
-        assert "creation_info" in attrs
-
-        creation_info = attrs["creation_info"]
-        assert creation_info["fits_extension"] == 0
-        assert creation_info["recursive_scan"] is False
-
-
-class TestNCHWFormat:
-    """Test that all images are converted to NCHW format correctly."""
-
-    def test_grayscale_to_nchw(self, temp_dir):
-        """Test grayscale image conversion to NCHW format."""
-        from images_to_zarr.convert import _ensure_nchw_format
-
-        # Create a simple grayscale image (H, W)
-        grayscale_data = np.random.randint(0, 255, (64, 64), dtype=np.uint8)
-
-        # Test _ensure_nchw_format directly
-        nchw_data = _ensure_nchw_format(grayscale_data)
-
-        # Should be (1, 1, 64, 64) - batch=1, channels=1, height=64, width=64
-        assert nchw_data.shape == (1, 1, 64, 64)
-        assert np.array_equal(nchw_data[0, 0, :, :], grayscale_data)
-
-    def test_rgb_hwc_to_nchw(self, temp_dir):
-        """Test RGB image in HWC format conversion to NCHW."""
-        from images_to_zarr.convert import _ensure_nchw_format
-
-        # Create RGB image in HWC format (Height, Width, Channels)
-        rgb_hwc = np.random.randint(0, 255, (64, 64, 3), dtype=np.uint8)
-
-        nchw_data = _ensure_nchw_format(rgb_hwc)
-
-        # Should be (1, 3, 64, 64) - batch=1, channels=3, height=64, width=64
-        assert nchw_data.shape == (1, 3, 64, 64)
-
-        # Check that data is correctly transposed
-        for c in range(3):
-            assert np.array_equal(nchw_data[0, c, :, :], rgb_hwc[:, :, c])
-
-    def test_fits_chw_to_nchw(self, temp_dir):
-        """Test FITS image in CHW format conversion to NCHW."""
-        from images_to_zarr.convert import _ensure_nchw_format
-
-        # Create FITS-style image in CHW format (Channels, Height, Width)
-        fits_chw = np.random.random((2, 64, 64)).astype(np.float32)
-
-        nchw_data = _ensure_nchw_format(fits_chw)
-
-        # Should be (1, 2, 64, 64) - batch=1, channels=2, height=64, width=64
-        assert nchw_data.shape == (1, 2, 64, 64)
-        assert np.array_equal(nchw_data[0, :, :, :], fits_chw)
-
-    def test_different_formats_produce_nchw(self, sample_images):
-        """Test that different image formats all produce NCHW output."""
-        from images_to_zarr.convert import _read_image_data, _ensure_nchw_format
-
-        images_dir, files = sample_images
-
-        for file_path in files:
-            # Read raw data and convert to NCHW
-            raw_data, metadata = _read_image_data(file_path)
-            data = _ensure_nchw_format(raw_data)
-
-            # All images should be in NCHW format (4D)
-            assert data.ndim == 4, f"Image {file_path.name} is not 4D: {data.shape}"
-            assert data.shape[0] == 1, f"Batch dimension should be 1: {data.shape}"
-
-            # Check that channels, height, width are positive
-            _, c, h, w = data.shape
-            assert c > 0, f"Channels dimension invalid: {c}"
-            assert h > 0, f"Height dimension invalid: {h}"
-            assert w > 0, f"Width dimension invalid: {w}"
-
-    def test_zarr_store_has_nchw_format(self, temp_dir, sample_images, sample_metadata):
-        """Test that the final Zarr store contains data in NCHW format."""
-        images_dir, files = sample_images
-        metadata_path, metadata_df = sample_metadata
-
-        # Convert to Zarr
-        zarr_path = convert(
-            output_dir=temp_dir,
-            folders=[images_dir],
-            metadata=metadata_path,
-            chunk_shape=(1, 128, 128),
-            overwrite=True,
-        )
-
-        # Open the Zarr store and check format
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]  # For grayscale images, should be 3D: (N, H, W)
-        # For multi-channel images, should be 4D: (N, C, H, W)
-        assert images_array.ndim in [3, 4], f"Zarr array should be 3D or 4D: {images_array.shape}"
-
-        # Check that we have the expected number of images
-        assert images_array.shape[0] == len(files)
-
-        if images_array.ndim == 4:
-            # Multi-channel format (N, C, H, W)
-            n, c, h, w = images_array.shape
-            assert c > 0 and h > 0 and w > 0
-        else:
-            # Grayscale format (N, H, W)
-            n, h, w = images_array.shape
-            assert h > 0 and w > 0
-
-
-class TestFoldersInputNormalization:
-    """Test that single string folders input is converted to list."""
-
-    def test_single_string_folder(self, temp_dir, sample_images):
-        """Test that a single string folder is converted to a list."""
-        images_dir, files = sample_images
-
-        # Test with single string
-        zarr_path = convert(
-            output_dir=temp_dir,
-            folders=str(images_dir),  # Single string, not list
-            overwrite=True,
-        )
-
-        # Should work and create a zarr store
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-
-    def test_single_path_folder(self, temp_dir, sample_images):
-        """Test that a single Path folder is converted to a list."""
-        images_dir, files = sample_images
-
-        # Test with single Path object
-        zarr_path = convert(
-            output_dir=temp_dir,
-            folders=images_dir,  # Single Path, not list
-            overwrite=True,
-        )
-
-        # Should work and create a zarr store
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-
-    def test_list_of_folders(self, temp_dir, sample_images):
-        """Test that a list of folders works correctly."""
-        images_dir, files = sample_images
-
-        # Create another folder with one image
-        images_dir2 = temp_dir / "images2"
-        images_dir2.mkdir()
-        sample_data = np.random.randint(0, 255, (32, 32), dtype=np.uint8)
-        Image.fromarray(sample_data, mode="L").save(images_dir2 / "extra.png")
-
-        # Test with list of folders
-        zarr_path = convert(
-            output_dir=temp_dir,
-            folders=[images_dir, images_dir2],  # List of folders
-            overwrite=True,
-        )
-
-        # Should work and create a zarr store with images from both folders
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-
-        # Check that we have images from both folders
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        assert images_array.shape[0] == len(files) + 1  # Original files + 1 extra
-
-
-class TestDirectImageConversion:
-    """Test converting images directly from memory."""
-
-    def test_convert_nchw(self, temp_dir):
-        """Test converting images directly from memory in NCHW format."""
-        # Create sample images in NCHW format
-        batch_size = 5
-        channels = 3
-        height = 64
-        width = 64
-
-        images = np.random.randint(0, 255, (batch_size, channels, height, width), dtype=np.uint8)
-
-        # Create metadata
-        metadata = [{"filename": f"memory_image_{i}.unknown", "id": i} for i in range(batch_size)]
-
-        # Convert from memory
-        zarr_path = convert(
-            output_dir=temp_dir,
-            images=images,
-            image_metadata=metadata,
-            overwrite=True,
-        )
-
-        # Check the result
-        assert zarr_path.exists()
-        assert zarr_path.is_dir()
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        # Should have the same shape and data
-        assert images_array.shape == images.shape
-        assert np.array_equal(images_array[:], images)
-
-    def test_convert_with_convenience_function(self, temp_dir):
-        """Test the convenience function convert."""
-
-        # Create sample images
-        images = np.random.random((3, 2, 32, 32)).astype(np.float32)
-
-        zarr_path = convert(
-            images=images,
-            output_dir=temp_dir,
-            overwrite=True,
-        )
-
-        assert zarr_path.exists()
-
-        # Check that the path structure is correct
-        assert zarr_path.name == "images.zarr"
-        assert zarr_path.parent == temp_dir
-
-        # Check that there's no nested images.zarr/images.zarr
-        nested_zarr = zarr_path / "images.zarr"
-        assert not nested_zarr.exists(), f"Found nested zarr structure: {nested_zarr}"
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        assert np.allclose(images_array[:], images)
-
-    def test_invalid_direct_image_input(self, temp_dir):
-        """Test that invalid direct image input raises appropriate errors."""
-        # Test with wrong number of dimensions
-        with pytest.raises(ValueError, match="Direct image input must be 4D"):
-            convert(
-                output_dir=temp_dir,
-                images=np.random.random((64, 64)),  # 2D instead of 4D
-                overwrite=True,
-            )
-
-        # Test with wrong number of dimensions
-        with pytest.raises(ValueError, match="Direct image input must be 4D"):
-            convert(
-                output_dir=temp_dir,
-                images=np.random.random((5, 64, 64)),  # 3D instead of 4D
-                overwrite=True,
-            )
-
-    def test_no_folders_and_no_images_error(self, temp_dir):
-        """Test that providing neither folders nor images raises an error."""
-        with pytest.raises(ValueError, match="Must provide either folders or images"):
-            convert(
-                output_dir=temp_dir,
-                folders=None,
-                images=None,
-                overwrite=True,
-            )
-
-
-class TestPathStructure:
-    """Test that zarr paths are created correctly for both folder and memory conversion."""
-
-    def test_path_structure_correctness(self, temp_dir, sample_images):
-        """Test that zarr paths are created correctly for both folder and memory conversion."""
-        # Test folder-based conversion
-        images_dir, files = sample_images
-        from images_to_zarr.convert import convert
-
-        zarr_path_folder = convert(
-            folders=[images_dir],
-            output_dir=temp_dir,
-            overwrite=True,
-        )
-
-        assert zarr_path_folder.exists()
-        assert zarr_path_folder.name == "images.zarr"
-        assert zarr_path_folder.parent == temp_dir
-
-        # Check that there's no nested zarr structure
-        nested_zarr_folder = zarr_path_folder / "images.zarr"
-        assert (
-            not nested_zarr_folder.exists()
-        ), f"Found nested zarr in folder conversion: {nested_zarr_folder}"
-
-        # Clean up
-        import shutil
-
-        shutil.rmtree(zarr_path_folder)
-
-        images = np.random.random((2, 3, 32, 32)).astype(np.float32)
-
-        zarr_path_memory = convert(
-            images=images,
-            output_dir=temp_dir,
-            overwrite=True,
-        )
-
-        assert zarr_path_memory.exists()
-        assert zarr_path_memory.name == "images.zarr"
-        assert zarr_path_memory.parent == temp_dir
-
-        # Check that there's no nested zarr structure
-        nested_zarr_memory = zarr_path_memory / "images.zarr"
-        assert (
-            not nested_zarr_memory.exists()
-        ), f"Found nested zarr in memory conversion: {nested_zarr_memory}"
-
-
-class TestResizingFeatures:
-    """Test new functionality added to images_to_zarr."""
-
-    def test_resize_functionality(self, temp_dir):
-        """Test that resize parameter works correctly."""
-        # Create images with different sizes (only 2 to ensure both are sampled)
-        images_dir = temp_dir / "mixed_sizes"
-        images_dir.mkdir()
-
-        # Create just 2 images with different dimensions to ensure both are analyzed
-        sizes = [(32, 48), (64, 64)]
-        files = []
-
-        for i, (h, w) in enumerate(sizes):
-            data = np.random.randint(0, 255, (h, w), dtype=np.uint8)
-            img_path = images_dir / f"image_{i}.png"
-            Image.fromarray(data, mode="L").save(img_path)
-            files.append(img_path)
-
-        output_dir = temp_dir / "output"
-
-        # Test 1: Without resize, should fail for mismatched dimensions
-        with pytest.raises(ValueError, match="All images must have the same dimensions"):
-            convert(
-                folders=[images_dir],
-                output_dir=output_dir,
-                overwrite=True,
-            )
-
-        # Test 2: With resize, should work and resize all images to target size
-        target_size = (50, 60)  # (height, width)
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir,
-            resize=target_size,
-            overwrite=True,
-        )
-
-        # Verify the zarr array has the correct dimensions
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        assert images_array.shape == (len(files), target_size[0], target_size[1])
-        assert root.attrs["creation_info"]["resize"] == list(target_size)
-
-    def test_interpolation_order(self, temp_dir):
-        """Test different interpolation orders for resizing."""
-        # Create a small test image
-        images_dir = temp_dir / "test_interp"
-        images_dir.mkdir()
-
-        # Create a simple image with clear patterns
-        data = np.zeros((20, 20), dtype=np.uint8)
-        data[5:15, 5:15] = 255  # White square in center
-        img_path = images_dir / "test.png"
-        Image.fromarray(data, mode="L").save(img_path)
-
-        output_dir = temp_dir / "output"
-
-        # Test different interpolation orders
-        for order in [0, 1, 3]:  # Nearest, linear, cubic
-            zarr_path = convert(
-                folders=[images_dir],
-                output_dir=output_dir / f"order_{order}",
-                resize=(40, 40),  # Double the size
-                interpolation_order=order,
-                overwrite=True,
-            )
-
-            store = zarr.storage.LocalStore(zarr_path)
-            root = zarr.open_group(store=store, mode="r")
-            assert root.attrs["creation_info"]["interpolation_order"] == order
-
-        # Test invalid interpolation order
-        with pytest.raises(ValueError, match="interpolation_order must be between 0 and 5"):
-            convert(
-                folders=[images_dir],
-                output_dir=output_dir / "invalid",
-                resize=(40, 40),
-                interpolation_order=10,  # Invalid
-                overwrite=True,
-            )
-
-
-class TestChunkShapeHandling:
-    """Test chunk_shape parameter handling."""
-
-    def test_chunk_shape_handling(self, temp_dir, sample_images):
-        """Test that chunk_shape parameter is handled correctly."""
-        images_dir, files = sample_images
-        output_dir = temp_dir / "output"
-
-        # Test 1: User-specified 3D chunk shape (should work for 3D arrays)
-        user_chunk_3d = (2, 32, 32)
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir / "chunk_3d",
-            chunk_shape=user_chunk_3d,
-            overwrite=True,
-        )
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        # Should respect user input (clamped to array size)
-        expected_chunks = tuple(min(c, s) for c, s in zip(user_chunk_3d, images_array.shape))
-        assert images_array.chunks == expected_chunks
-
-        # Test 2: User-specified chunk shape that's too large (should be clamped)
-        large_chunk = (100, 200, 200)  # Larger than image dimensions
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir / "chunk_large",
-            chunk_shape=large_chunk,
-            overwrite=True,
-        )
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        # Should be clamped to actual array size
-        expected_chunks = tuple(min(c, s) for c, s in zip(large_chunk, images_array.shape))
-        assert images_array.chunks == expected_chunks
-
-    def test_chunk_shape_with_channels(self, temp_dir):
-        """Test chunk_shape handling with multi-channel images."""
-        # Create RGB images
-        images_dir = temp_dir / "rgb_images"
-        images_dir.mkdir()
-
-        # Create 3-channel RGB images
-        for i in range(3):
-            data = np.random.randint(0, 255, (64, 64, 3), dtype=np.uint8)
-            img_path = images_dir / f"rgb_{i}.png"
-            Image.fromarray(data, mode="RGB").save(img_path)
-
-        output_dir = temp_dir / "output"
-
-        # Test 3D chunk shape with 4D array (should expand to include channels)
-        user_chunk_3d = (1, 32, 32)
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir,
-            chunk_shape=user_chunk_3d,
-            overwrite=True,
-        )
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        # Should be 4D array with channels
-        assert len(images_array.shape) == 4  # (N, C, H, W)
-        assert images_array.shape[1] == 3  # 3 channels
-        # Chunk shape should be expanded to include channels
-        expected_chunks = (1, 3, 32, 32)  # Full channels dimension
-        assert images_array.chunks == expected_chunks
-
-
-class TestDisplay:
-    """Test the display_sample_images function."""
-
-    def test_display_sample_images(self, temp_dir, sample_images):
-        """Test the display_sample_images function."""
-        from images_to_zarr.display_sample_images import display_sample_images
-
-        images_dir, files = sample_images
-        output_dir = temp_dir / "output"
-
-        # Create a zarr store
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir,
-            overwrite=True,
-        )
-        # Test basic display (should not raise errors)
-        try:
-            # Test with default parameters (save to file to avoid blocking during tests)
-            save_path1 = temp_dir / "test_display_default.png"
-            display_sample_images(zarr_path, num_samples=2, figsize=(8, 6), save_path=save_path1)
-
-            # Test with saving to file
-            save_path2 = temp_dir / "test_display.png"
-            display_sample_images(zarr_path, num_samples=1, save_path=save_path2)
-
-            # Test with all images if fewer than num_samples (save to file)
-            save_path3 = temp_dir / "test_display_many.png"
-            display_sample_images(
-                zarr_path, num_samples=10, save_path=save_path3
-            )  # More than available
-
-        except ImportError:
-            # matplotlib not available - this is expected in some test environments
-            pytest.skip("matplotlib not available for display testing")
-        except Exception as e:
-            # The function might fail in a headless environment, but shouldn't crash
-            # due to missing display. We mainly want to test the data loading logic.
-            if "display" not in str(e).lower() and "DISPLAY" not in str(e):
-                raise e
-
-    def test_display_with_different_dtypes(self, temp_dir):
-        """Test display_sample_images with different data types and ranges."""
-        from images_to_zarr.display_sample_images import display_sample_images
-
-        # Create images with different dtypes
-        images_dir = temp_dir / "dtype_test"
-        images_dir.mkdir()
-
-        # Create float32 FITS image with values in range [0, 1]
-        data_float = np.random.random((64, 64)).astype(np.float32)
-        fits_path = images_dir / "float_image.fits"
-        hdu = fits.PrimaryHDU(data_float)
-        hdu.writeto(fits_path, overwrite=True)
-
-        # Create uint16 image
-        data_uint16 = (np.random.random((64, 64)) * 65535).astype(np.uint16)
-        fits_path_16 = images_dir / "uint16_image.fits"
-        hdu16 = fits.PrimaryHDU(data_uint16)
-        hdu16.writeto(fits_path_16, overwrite=True)
-
-        output_dir = temp_dir / "output"
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=output_dir,
-            overwrite=True,
-        )
-
-        # Test display with auto-normalization
-        try:
-            save_path = temp_dir / "dtype_display.png"
-            display_sample_images(zarr_path, num_samples=2, save_path=save_path)
-        except ImportError:
-            pytest.skip("matplotlib not available for display testing")
-        except Exception as e:
-            # The function might fail in a headless environment
-            if "display" not in str(e).lower() and "DISPLAY" not in str(e):
-                raise e
-
-    def test_path_handling(self, temp_dir, sample_images):
-        """Test correct path handling for output directories."""
-        images_dir, files = sample_images
-
-        # Test 1: Output path ending with .zarr should be used directly
-        zarr_output = temp_dir / "custom_name.zarr"
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=zarr_output,
-            overwrite=True,
-        )
-
-        assert zarr_path == zarr_output
-        assert zarr_path.exists()
-
-        # Test 2: Output path not ending with .zarr should create .zarr inside
-        regular_output = temp_dir / "regular_dir"
-        zarr_path = convert(
-            folders=[images_dir],
-            output_dir=regular_output,
-            overwrite=True,
-        )
-
-        assert zarr_path.parent == regular_output
-        assert zarr_path.name == "images.zarr"
-        assert zarr_path.exists()
-
-        # Test 3: With metadata file, should use metadata filename
-        metadata_data = [{"filename": f.name} for f in files]
-        metadata_df = pd.DataFrame(metadata_data)
-        metadata_path = temp_dir / "my_dataset.csv"
-        metadata_df.to_csv(metadata_path, index=False)
-
-        zarr_path = convert(
-            folders=[images_dir],
-            metadata=metadata_path,
-            output_dir=regular_output / "with_metadata",
-            overwrite=True,
-        )
-
-        assert zarr_path.name == "my_dataset.zarr"
-
-
-class TestErrorHandling:
-    """Test error handling for edge cases in new functionality."""
-
-    def test_error_handling_edge_cases(self, temp_dir):
-        """Test error handling for edge cases in new functionality."""
-        output_dir = temp_dir / "output"
-
-        # Test 1: No folders and no images provided
-        with pytest.raises(ValueError, match="Must provide either folders or images"):
-            convert(output_dir=output_dir)
-
-        # Test 2: Invalid images array for direct conversion
-        with pytest.raises(ValueError, match="images parameter must be a numpy array"):
-            convert(images="not_an_array", output_dir=output_dir)
-
-        # Test 3: Wrong dimensionality for direct images
-        with pytest.raises(ValueError, match="Direct image input must be 4D"):
-            convert(images=np.random.random((64, 64)), output_dir=output_dir)
-
-        # Test 4: Empty folder (no images found)
-        empty_dir = temp_dir / "empty"
-        empty_dir.mkdir()
-
-        with pytest.raises(ValueError, match="No image files found"):
-            convert(folders=[empty_dir], output_dir=output_dir)
-
-    def test_direct_memory_conversion(self, temp_dir):
-        """Test conversion from numpy arrays directly."""
-        # Create test images in memory
-        images = np.random.randint(0, 255, (5, 1, 64, 64), dtype=np.uint8)
-
-        # Create corresponding metadata
-        metadata = [
-            {
-                "original_filename": f"memory_image_{i}.png",
-                "dtype": str(images.dtype),
-                "shape": images.shape[1:],
-                "custom_field": f"value_{i}",
-            }
-            for i in range(len(images))
-        ]
-
-        output_dir = temp_dir / "output"
-        zarr_path = convert(
-            images=images,
-            image_metadata=metadata,
-            output_dir=output_dir,
-            chunk_shape=(2, 1, 32, 32),  # Test custom chunking with correct 4D shape
-            overwrite=True,
-        )
-
-        # Verify the conversion
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        assert images_array.shape == images.shape
-        assert np.array_equal(images_array[:], images)
-        assert root.attrs["creation_info"]["direct_memory_conversion"] is True
-
-        # Check metadata
-        metadata_path = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        assert metadata_path.exists()
-
-        loaded_metadata = pd.read_parquet(metadata_path)
-        assert len(loaded_metadata) == len(metadata)
-        assert "custom_field" in loaded_metadata.columns
-
-
-class TestComprehensiveIntegration:
-    """Test a comprehensive integration of all features."""
-
-    def test_comprehensive_integration(self, temp_dir):
-        """Test a comprehensive scenario with multiple features."""
-        # Create a complex scenario with:
-        # - Mixed image formats and sizes
-        # - Custom metadata
-        # - Resize functionality
-        # - Custom chunking
-        # - Custom compression
-
-        images_dir = temp_dir / "complex_test"
-        images_dir.mkdir()
-
-        # Create images with different formats and sizes
-        formats_and_sizes = [
-            ("png", (32, 48), np.uint8),
-            ("jpg", (64, 64), np.uint8),
-            ("tiff", (100, 80), np.uint8),
-        ]
-
-        files = []
-        metadata_entries = []
-
-        for i, (fmt, (h, w), dtype) in enumerate(formats_and_sizes):
-            # Create random image data
-            if fmt == "jpg":
-                # JPEG needs RGB
-                data = np.random.randint(0, 255, (h, w, 3), dtype=dtype)
-                mode = "RGB"
-            else:
-                data = np.random.randint(0, 255, (h, w), dtype=dtype)
-                mode = "L"
-
-            img_path = images_dir / f"complex_{i}.{fmt}"
-            Image.fromarray(data, mode=mode).save(img_path)
-            files.append(img_path)
-
-            # Create metadata
-            metadata_entries.append(
-                {
-                    "filename": img_path.name,
-                    "object_id": f"OBJ_{i:03d}",
-                    "ra": 180.0 + i * 10.0,
-                    "dec": -30.0 + i * 5.0,
-                    "filter": ["g", "r", "i"][i],
-                    "exposure_time": [30, 60, 120][i],
-                }
-            )
-
-        # Create metadata file
-        metadata_df = pd.DataFrame(metadata_entries)
-        metadata_path = temp_dir / "complex_metadata.csv"
-        metadata_df.to_csv(metadata_path, index=False)
-
-        # Run conversion with all features
-        output_dir = temp_dir / "complex_output.zarr"
-        zarr_path = convert(
-            folders=[images_dir],
-            metadata=metadata_path,
-            output_dir=output_dir,
-            resize=(50, 60),  # Resize all to same size
-            interpolation_order=1,  # Bilinear
-            chunk_shape=(2, 25, 30),  # Custom chunk
-            compressor="zstd",
-            clevel=2,
-            num_parallel_workers=2,
-            overwrite=True,
-        )
-
-        # Verify comprehensive results
-        assert zarr_path.exists()
-        assert zarr_path == output_dir  # Used exact path
-
-        store = zarr.storage.LocalStore(zarr_path)
-        root = zarr.open_group(store=store, mode="r")
-        images_array = root["images"]
-
-        # Check array properties
-        assert images_array.shape == (3, 3, 50, 60)  # 3 images, 3 channels (RGB), resized
-        assert images_array.chunks == (2, 3, 25, 30)  # Custom chunk + full channels
-
-        # Check attributes
-        attrs = root.attrs
-        assert attrs["compressor"] == "zstd"
-        assert attrs["compression_level"] == 2
-        assert attrs["creation_info"]["resize"] == [50, 60]
-        assert attrs["creation_info"]["interpolation_order"] == 1
-
-        # Check metadata preservation
-        metadata_path_out = zarr_path.parent / f"{zarr_path.stem}_metadata.parquet"
-        combined_metadata = pd.read_parquet(metadata_path_out)
-
-        assert len(combined_metadata) == 3
-        assert "object_id" in combined_metadata.columns
-        assert "filter" in combined_metadata.columns
-        assert "original_filename" in combined_metadata.columns  # Added by processing
-
-
-if __name__ == "__main__":
-    pytest.main([__file__])
